docs/template controller 파일 정리
